api/requirements.txt
Flask
Flask-Cors

Werkzeug
PyJWT
firebase-admin